fixing broken script engine code

+++ elenasrc2/elenasm/scriptengine.h
@@ -134,8 +134,8 @@ struct ScriptBookmark
 {
    pos_t offset;
    int   state;
-   short row;
-   short column;
+   int   row;
+   int   column;
 
    ScriptBookmark()
    {
